+++ src/main/java/roomescape/domain/reservation/controller/AdminReservationController.java
@@ -11,9 +11,9 @@ import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.login.controller.MemberResolver;
 import roomescape.domain.member.domain.Member;
 import roomescape.domain.reservation.domain.reservation.Reservation;
-import roomescape.domain.reservation.dto.ReservationAddRequest;
-import roomescape.domain.reservation.dto.ReservationFindRequest;
-import roomescape.domain.reservation.dto.ReservationResponse;
+import roomescape.domain.reservation.dto.request.ReservationAddRequest;
+import roomescape.domain.reservation.dto.request.ReservationFindRequest;
+import roomescape.domain.reservation.dto.response.ReservationResponse;
 import roomescape.domain.reservation.service.ReservationService;
 
 @RestController
