+++ src/main/java/roomescape/domain/reservation/controller/AdminReservationTimeController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.reservation.domain.reservationTime.ReservationTime;
-import roomescape.domain.reservation.dto.ReservationTimeAddRequest;
+import roomescape.domain.reservation.dto.request.ReservationTimeAddRequest;
 import roomescape.domain.reservation.service.AdminReservationTimeService;
 
 @RestController
