+++ src/main/java/roomescape/domain/reservation/dto/request/BookableTimesRequest.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservation.dto;
+package roomescape.domain.reservation.dto.request;
 
 import java.time.LocalDate;
