+++ src/main/java/roomescape/domain/reservation/dto/response/ReservationMineResponse.java
@@ -1,7 +1,8 @@
-package roomescape.domain.reservation.dto;
+package roomescape.domain.reservation.dto.response;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
+import roomescape.domain.reservation.dto.ReservationWithOrderDto;
 
 public record ReservationMineResponse(Long reservationId, String theme, LocalDate date, LocalTime time, String status) {
