+++ src/main/java/roomescape/domain/reservation/dto/response/ReservationResponse.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservation.dto;
+package roomescape.domain.reservation.dto.response;
 
 import java.time.LocalDate;
 import java.util.List;
