+++ src/main/java/roomescape/domain/reservation/service/AdminReservationTimeService.java
@@ -3,7 +3,7 @@ package roomescape.domain.reservation.service;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.domain.reservation.domain.reservationTime.ReservationTime;
-import roomescape.domain.reservation.dto.ReservationTimeAddRequest;
+import roomescape.domain.reservation.dto.request.ReservationTimeAddRequest;
 import roomescape.domain.reservation.repository.reservationTime.ReservationTimeRepository;
 import roomescape.global.exception.EscapeApplicationException;
 import roomescape.global.exception.NoMatchingDataException;
