+++ src/main/java/roomescape/domain/reservation/service/ReservationService.java
@@ -12,10 +12,10 @@ import roomescape.domain.member.repository.MemberRepository;
 import roomescape.domain.reservation.domain.reservation.Reservation;
 import roomescape.domain.reservation.domain.reservation.ReservationStatus;
 import roomescape.domain.reservation.domain.reservationTime.ReservationTime;
-import roomescape.domain.reservation.dto.BookableTimeResponse;
-import roomescape.domain.reservation.dto.BookableTimesRequest;
-import roomescape.domain.reservation.dto.ReservationAddRequest;
-import roomescape.domain.reservation.dto.ReservationMineResponse;
+import roomescape.domain.reservation.dto.request.BookableTimesRequest;
+import roomescape.domain.reservation.dto.request.ReservationAddRequest;
+import roomescape.domain.reservation.dto.response.BookableTimeResponse;
+import roomescape.domain.reservation.dto.response.ReservationMineResponse;
 import roomescape.domain.reservation.repository.reservation.ReservationRepository;
 import roomescape.domain.reservation.repository.reservationTime.ReservationTimeRepository;
 import roomescape.domain.theme.domain.Theme;
