+++ src/test/java/roomescape/domain/reservation/controller/ReservationControllerTest.java
@@ -15,7 +15,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import roomescape.ControllerTest;
-import roomescape.domain.reservation.dto.ReservationAddRequest;
+import roomescape.domain.reservation.dto.request.ReservationAddRequest;
 
 class ReservationControllerTest extends ControllerTest {
