+++ src/test/java/roomescape/domain/reservation/service/ReservationServiceTest.java
@@ -28,10 +28,10 @@ import roomescape.domain.member.domain.Member;
 import roomescape.domain.member.service.FakeMemberRepository;
 import roomescape.domain.reservation.domain.reservation.Reservation;
 import roomescape.domain.reservation.domain.reservationTime.ReservationTime;
-import roomescape.domain.reservation.dto.BookableTimeResponse;
-import roomescape.domain.reservation.dto.BookableTimesRequest;
-import roomescape.domain.reservation.dto.ReservationAddRequest;
-import roomescape.domain.reservation.dto.ReservationMineResponse;
+import roomescape.domain.reservation.dto.request.BookableTimesRequest;
+import roomescape.domain.reservation.dto.request.ReservationAddRequest;
+import roomescape.domain.reservation.dto.response.BookableTimeResponse;
+import roomescape.domain.reservation.dto.response.ReservationMineResponse;
 import roomescape.domain.theme.domain.Theme;
 import roomescape.domain.theme.service.FakeThemeRepository;
 import roomescape.global.exception.EscapeApplicationException;
